Preparing for stabilization of fN::BITS

+++ src/impls.rs
@@ -980,7 +980,8 @@ macro_rules! impl_float {
     ($($ty:ty, $aty:ty, $zero:expr, $one:expr,)*) => {$(
 
 impl AsBytes for $aty {
-    const BITS: usize = <$ty>::BITS;
+    #[expect(clippy::unnecessary_cast, reason = "Preparing for stabilization of fN::BITS.")]
+    const BITS: usize = <$ty>::BITS as usize;
     const BYTES: usize = <$ty>::BYTES;
     type Bytes = [u8;  <$ty>::BYTES];
 }
